Verhindern, dass alle Profile auf einmal gelöscht werden können
closes #1508

- Wenn kein Profil ausgewählt ist dann nichts löschen
- Wenn die zu löschende Node selbst Kinder hat nichts tun, damit nur die
"untersten", also die wirklichen Workspaces, entfernt werden können

Außerdem Logging einbauen

+++ bundles/aero.minova.rcp.workspace/src/aero/minova/rcp/workspace/dialogs/WorkspaceDialog.java
@@ -122,7 +122,15 @@ public class WorkspaceDialog extends Dialog {
 		deleteProfile.addMouseListener(new MouseAdapter() {
 			@Override
 			public void mouseDown(MouseEvent e) {
-				WorkspaceAccessPreferences.deleteSavedWorkspace(profile.getText());
+
+				String profileNameToDelete = profile.getText();
+				logger.info("Deleting Profile \"" + profileNameToDelete + "\"");
+				if (profileNameToDelete == null || profileNameToDelete.isBlank()) {
+					logger.info("No profile selected, not deleting");
+					return;
+				}
+
+				WorkspaceAccessPreferences.deleteSavedWorkspace(profileNameToDelete);
 				try {
 					// Workspace Ordner löschen
 					FileUtils.deleteDirectory(new File(applicationArea.getText().replace("file:", "")));
